game_app/requirements.txt
# Web framework
Django==5.1.2
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Cache
redis==5.0.8

# Background tasks
celery==5.4.0

# CORS for API
django-cors-headers==4.4.0

# Static files
whitenoise==6.7.0

# HTTP requests
requests==2.32.3

# Environment variables
python-decouple==3.8

# Development
django-debug-toolbar==4.4.2

# Production monitoring
sentry-sdk==2.10.0
